refactor: user controllers refactored
