+++ backend/src/controllers/user/index.ts
@@ -1,0 +1,1 @@
+export { default as getProfileInformation } from './getProfileInfo/getProfileInfo';
